backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "pytest>=9.0.1",
    "uvicorn>=0.38.0",
]
